+++ requirements.txt
@@ -6,3 +6,4 @@ pillow==8.1.0
 PyMySQL==1.0.2
 pathlib==1.0.1
 email-validator==1.1.2
+flask-paginate==0.8.1
