+++ templates/admin/admin.html
@@ -4,9 +4,9 @@
         <h3>Tableau d'administration</h3>
         <p class='text-muted'>
             <ul class="list-group">
-                <a class="btn" href="{{ url_for('admin_users') }}"><li class="list-group-item list-group-item-light">Utilisateurs</li></a>
-                <a class="btn" href="{{ url_for('admin_posts') }}"><li class="list-group-item list-group-item-light">Posts</li></a>
-                <a class="btn" href="{{ url_for('logs') }}"><li class="list-group-item list-group-item-light">Logs</li></a>
+                <a class="btn" href="{{ url_for('admin_user_page.admin_users') }}"><li class="list-group-item list-group-item-light">Utilisateurs</li></a>
+                <a class="btn" href="{{ url_for('admin_posts_page.admin_posts') }}"><li class="list-group-item list-group-item-light">Posts</li></a>
+                <a class="btn" href="{{ url_for('admin_logs_page.logs') }}"><li class="list-group-item list-group-item-light">Logs</li></a>
             </ul>
         </p>
     </div>
