+++ templates/admin/users.html
@@ -7,8 +7,7 @@
             <th scope="col">Permission</th>
             <th scope="col">Username</th>
             <th scope="col">Email</th>
-            <th scope="col">Edit</th>
-            <th scope="col">Delete</th>
+            <th scope="col">Actions</th>
         </tr>
         </thead>
         <tbody>
@@ -18,8 +17,7 @@
                 <td>{{ user['permission'] }}</td>
                 <td>{{ user['username']}}</td>
                 <td>{{ user['email']}}</td>
-                <td><a class="btn btn-secondary" href="{{ url_for('admin_edit_user', username = user['username']) }}">Edit</a></td>
-                <td><a class="btn btn-danger" href="{{ url_for('admin_delete_user', username = user['username']) }}">Delete</a></td>
+                <td><a class="btn btn-secondary mr-1" href="{{ url_for('admin_user_page.admin_edit_user', username = user['username']) }}">Edit</a><a class="btn btn-danger ml-1" href="{{ url_for('admin_user_page.admin_delete_user', username = user['username']) }}">Delete</a></td>
             </tr>
         {% endfor %}
         </tbody>
