+++ templates/authentification/login.html
@@ -47,7 +47,7 @@
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Need an account ? <a class="ml-2" href="{{ url_for('register') }}">Sign up now !</a>
+            Need an account ? <a class="ml-2" href="{{ url_for('registration.register') }}">Sign up now !</a>
         </small>
     </div>
 {% endblock content %}
