+++ templates/error/404.html
@@ -9,7 +9,7 @@
         <p>Cette page n'existe pas</p>
         <p class='text-muted'>
             <ul class="list-group">
-                <a class="btn btn-block" href="{{ url_for('home') }}"><li class="list-group-item list-group-item-light">Page d'accueil</li></a>
+                <a class="btn btn-block" href="{{ url_for('home_page.home') }}"><li class="list-group-item list-group-item-light">Page d'accueil</li></a>
             </ul>
         </p>
     </div>
