+++ templates/register/register.html
@@ -140,7 +140,7 @@
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Vous avez déjà un compte ? <a class="ml-2" href="{{ url_for('login') }}">Connectez vous !</a>
+            Vous avez déjà un compte ? <a class="ml-2" href="{{ url_for('authentification.login') }}">Connectez vous !</a>
         </small>
     </div>
 {% endblock content %}
